pyproject.toml
[project]
name = "project-may"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "Django==5.2.8",
  "google-genai>=1.66.0",
  "pillow==12.0.0",
  "psycopg2-binary==2.9.11",
  "python-dotenv==1.2.1",
]

[dependency-groups]
dev = [
  "django-stubs>=5.2.9",
  "django-stubs-ext>=5.2.9",
  "ruff==0.15.1",
]

[tool.ruff]
line-length = 100

[tool.ruff.format]
quote-style = "single"
indent-style = "space"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]
